build: downgrade to Roslyn to 3.7.0
