+++ code/src/F0.CodeAnalysis.CSharp.Analyzer.Benchmarking/F0.CodeAnalysis.CSharp.Analyzer.Benchmarking.csproj
@@ -15,7 +15,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.2.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.7.0" />
   </ItemGroup>
 
   <ItemGroup>
