+++ code/src/F0.CodeAnalysis.CSharp.Benchmarking/F0.CodeAnalysis.CSharp.Benchmarking.csproj
@@ -16,7 +16,7 @@
 
   <ItemGroup>
     <PackageReference Include="DiffPlex" Version="1.7.1" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.8.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.7.0" />
   </ItemGroup>
 
 </Project>
